+++ etc/ui-vue-docs/.vitepress/app-components/AppPlayControlsContainer.vue
@@ -1,0 +1,6 @@
+<template>
+  <div class="app-play-container pa-8 mt-2">
+    <div class="mb-9 mt-9 text-mono-m">Try it</div>
+    <slot />
+  </div>
+</template>