+++ etc/ui-vue-docs/.vitepress/app-components/AppSubheader.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="text-subtitle-m mb-4">
+    <slot />
+  </div>
+</template>