+++ sdk/model/src/components/PlDataTable.ts
@@ -1,4 +1,15 @@
-import { JoinEntry, PColumnIdAndSpec, PTableRecordFilter, PTableSorting } from '@milaboratories/pl-model-common';
+import {
+  getAxisId,
+  isPColumn,
+  JoinEntry,
+  matchAxisId,
+  PColumn,
+  PColumnIdAndSpec,
+  PTableHandle,
+  PTableRecordFilter,
+  PTableSorting
+} from '@milaboratories/pl-model-common';
+import { RenderCtx, TreeNodeAccessor } from '../render';
 
 /** Data table state */
 export type PlDataTableGridState = {
@@ -42,3 +53,40 @@ export type PlDataTableState = {
   // mapping of gridState onto the p-table data structures
   pTableParams?: PTableParams;
 };
+
+/**
+ * Create p-table handle given ui table state
+ *
+ * @param ctx context
+ * @param columns column list
+ * @param tableState table ui state
+ * @returns
+ */
+export function createPlDataTable<A, U>(
+  ctx: RenderCtx<A, U>,
+  columns: PColumn<TreeNodeAccessor>[],
+  tableState?: PlDataTableState
+): PTableHandle {
+  const allLabelCols = ctx.resultPool
+    .getData()
+    .entries.map((d) => d.obj)
+    .filter(isPColumn)
+    .filter((p) => p.spec.name === 'pl7.app/label' && p.spec.axesSpec.length === 1);
+
+  const moreColumns = [];
+  for (const col of columns) {
+    for (const axis of col.spec.axesSpec) {
+      const axisId = getAxisId(axis);
+      for (const match of allLabelCols) {
+        if (matchAxisId(axisId, getAxisId(match.spec.axesSpec[0]))) {
+          moreColumns.push(match);
+        }
+      }
+    }
+  }
+  return ctx.createPTable({
+    columns: [...columns, ...moreColumns],
+    filters: tableState?.pTableParams?.filters,
+    sorting: tableState?.pTableParams?.sorting
+  });
+}