+++ sdk/ui-vue/src/components/PlAgDataTable/index.ts
@@ -1,3 +1,5 @@
 export { default as PlDataTable } from './PlAgDataTable.vue';
+export { default as PlAgOverlayLoading } from './PlAgOverlayLoading.vue';
+export { default as PlAgOverlayNoRows } from './PlAgOverlayNoRows.vue';
 
 export * from './types';